refactor: extract shared getSearchProjects and event labels into lib
Deduplicate getSearchProjects (was copy-pasted in 3 tool files) into
src/lib/search-projects.ts and TYPE_ICONS/TYPE_LABELS (duplicated in 2
files) into src/lib/event-labels.ts.

Updated: export-report, timeline-view, search-history to use shared modules.
Also derive TYPE_BADGES from TYPE_ICONS in search-history to stay DRY.
